+++ src/tests/test_vscode.py
@@ -2,12 +2,13 @@ import shutil
 import subprocess
 from pathlib import Path
 
+from _pytest.monkeypatch import MonkeyPatch
 from buildenv import BuildEnvManager
 from nmk.tests.tester import NmkBaseTester
 from nmk.utils import is_windows
 
 from nmk_vscode import __version__
-from nmk_vscode.buildenv import BuildEnvInit
+from nmk_vscode._buildenv.legacy import BuildEnvInit
 
 
 class TestVSCodePlugin(NmkBaseTester):
@@ -53,7 +54,7 @@ class TestVSCodePlugin(NmkBaseTester):
         )
         assert (self.test_folder / ".vscode" / "extensions.json").is_file()
 
-    def test_buildenv_extension(self, monkeypatch):
+    def test_buildenv_extension_legacy(self, monkeypatch: MonkeyPatch):
         # Prepare extension instance
         fake_venv_bin = self.test_folder / "venv" / ("Scripts" if is_windows() else "bin")
         if fake_venv_bin.is_dir():
@@ -68,20 +69,20 @@ class TestVSCodePlugin(NmkBaseTester):
         assert ext.get_version() == __version__
 
         # Check init without code command
-        monkeypatch.setattr(shutil, "which", lambda _: None)
+        monkeypatch.setattr(shutil, "which", lambda _: None)  # type: ignore
         ext.init(True)
         assert not activate_script.is_file()
         self.check_logs("nmk-vscode: 'code' command not found!")
-        monkeypatch.setattr(shutil, "which", lambda _: "code")
+        monkeypatch.setattr(shutil, "which", lambda _: "code")  # type: ignore
 
         # Check init with code returning an error
-        monkeypatch.setattr(subprocess, "run", lambda args, **kwargs: subprocess.CompletedProcess(args, 1))
+        monkeypatch.setattr(subprocess, "run", lambda args, **kwargs: subprocess.CompletedProcess(args, 1))  # type: ignore
         ext.init(True)
         assert not activate_script.is_file()
         self.check_logs("nmk-vscode: 'code --locate-shell-integration-path bash' command failed!")
 
         # Check init with code returning an invalid path
-        monkeypatch.setattr(subprocess, "run", lambda args, **kwargs: subprocess.CompletedProcess(args, 0, stdout="some/unknown/path/to/unknown/file"))
+        monkeypatch.setattr(subprocess, "run", lambda args, **kwargs: subprocess.CompletedProcess(args, 0, stdout="some/unknown/path/to/unknown/file"))  # type: ignore
         ext.init(True)
         assert not activate_script.is_file()
         self.check_logs("nmk-vscode: bash init script not found: ")
@@ -89,6 +90,6 @@ class TestVSCodePlugin(NmkBaseTester):
         # Check correct init
         fake_init = self.test_folder / "fake_vs_init.sh"
         fake_init.touch()
-        monkeypatch.setattr(subprocess, "run", lambda args, **kwargs: subprocess.CompletedProcess(args, 0, stdout=str(fake_init)))
+        monkeypatch.setattr(subprocess, "run", lambda args, **kwargs: subprocess.CompletedProcess(args, 0, stdout=str(fake_init)))  # type: ignore
         ext.init(True)
         assert activate_script.is_file()