+++ doc/changelog.md
@@ -8,6 +8,12 @@ Only interface and important behavior changes are listed here.
 The fully detailed changelog is also available on [Github](https://github.com/dynod/nmk-vscode/releases)
 ```
 
+## Release 1.2.0
+
+- Update for buildenv 2 handling:
+  - terminal integration in settings
+  - nmk tasks for setup/build/tests
+
 ## Release 1.1.0
 
 - Update generated settings: